Added Mass field and fixed wrong tire nodes calculation

+++ src/calculators/checks/PreChecks.java
@@ -44,11 +44,11 @@ public class PreChecks {
         int a = sumCheck.a();
         int h = sumCheck.h();
         int commonDivisor = 150 + t + a + h;
-        float finalT = (t*calculator.tValue0 + a*calculator.tValue1 + h*calculator.tValue2 + calculator.tValue3) / commonDivisor;
-        float finalA = (t*calculator.aValue0 + a*calculator.aValue1 + h*calculator.aValue2 + calculator.aValue3) / commonDivisor;
-        float finalH = (t*calculator.hValue0 + a*calculator.hValue1 + h*calculator.hValue2 + calculator.hValue3) / commonDivisor;
+        int finalT = (int) ((t*calculator.tValue0 + a*calculator.tValue1 + h*calculator.tValue2 + calculator.tValue3) / commonDivisor);
+        int finalA = (int) ((t*calculator.aValue0 + a*calculator.aValue1 + h*calculator.aValue2 + calculator.aValue3) / commonDivisor);
+        int finalH = (int) ((t*calculator.hValue0 + a*calculator.hValue1 + h*calculator.hValue2 + calculator.hValue3) / commonDivisor);
 
-        return (((int)finalT + (int)finalA + (int)finalH) / 3) > calculator.overall.max();
+        return ((finalT + finalA + finalH) / 3) > calculator.overall.max();
     }
 }
